# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: UDAS Member Management API
  description: API for managing organization members and application users
  version: v1

servers:
  - url: http://localhost:8080/api/v1
    description: Development server

paths:
  /members:
    post:
      summary: Create a new member
      operationId: createMember
      tags:
        - Members
      security:
        - oauth2: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Member'
      responses:
        '201':
          description: Member created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    get:
      summary: List all members with pagination and filtering
      operationId: listMembers
      tags:
        - Members
      security:
        - oauth2: []
      parameters:
        - name: firstName
          in: query
          schema:
            type: string
          description: Filter by first name
        - name: lastName
          in: query
          schema:
            type: string
          description: Filter by last name
        - name: email
          in: query
          schema:
            type: string
          description: Filter by email
        - name: phone
          in: query
          schema:
            type: string
          description: Filter by phone
        - name: city
          in: query
          schema:
            type: string
          description: Filter by city
        - name: country
          in: query
          schema:
            type: string
          description: Filter by country
        - name: page
          in: query
          schema:
            type: integer
            default: 0
          description: Page number
        - name: size
          in: query
          schema:
            type: integer
            default: 20
          description: Page size
        - name: sort
          in: query
          schema:
            type: string
            default: id,asc
          description: "Sort format: field,direction (e.g., lastName,asc)"
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedMember'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /members/{id}:
    get:
      summary: Get a member by ID
      operationId: getMember
      tags:
        - Members
      security:
        - oauth2: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '404':
          description: Member not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    put:
      summary: Update a member
      operationId: updateMember
      tags:
        - Members
      security:
        - oauth2: [READ_WRITE, ADMIN]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Member'
      responses:
        '200':
          description: Member updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '404':
          description: Member not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    delete:
      summary: Delete a member
      operationId: deleteMember
      tags:
        - Members
      security:
        - oauth2: [READ_WRITE, ADMIN]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Member deleted successfully
        '404':
          description: Member not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /users:
    post:
      summary: Create application user
      operationId: createUser
      tags:
        - Users
      security:
        - oauth2: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationUser'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    get:
      summary: List all users
      operationId: listUsers
      tags:
        - Users
      security:
        - oauth2: [ADMIN]
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedApplicationUser'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /users/{id}:
    get:
      summary: Get user by ID
      operationId: getUser
      tags:
        - Users
      security:
        - oauth2: [ADMIN]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationUser'
        '404':
          description: User not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    put:
      summary: Update user
      operationId: updateUser
      tags:
        - Users
      security:
        - oauth2: [ADMIN]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationUser'
        '404':
          description: User not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    delete:
      summary: Delete user
      operationId: deleteUser
      tags:
        - Users
      security:
        - oauth2: [ADMIN]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            read: Read access
            write: Write access

  schemas:
    Member:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: array
          items:
            type: string
            format: email
        phone:
          type: array
          items:
            type: string
        dateOfBirth:
          type: string
          format: date
        dateOfDeath:
          type: string
          format: date
        ssn:
          type: string
        address:
          $ref: '#/components/schemas/Address'

    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string

    PagedMember:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer

    ApplicationUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          readOnly: true
        name:
          type: string
        role:
          type: string
          enum: [READ_ONLY, READ_WRITE, ADMIN]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true

    ApplicationUserRequest:
      type: object
      required:
        - email
        - role
      properties:
        email:
          type: string
        role:
          type: string
          enum: [READ_ONLY, READ_WRITE, ADMIN]
        active:
          type: boolean
          default: true

    PagedApplicationUser:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationUser'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer

    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object